// src/Core/Common/GLAttributeInfo.d.ts
export default class GLAttributeInfo {
    attributeLocation: number;
    componentsNumberPerVertexAttribute: number;
    typeComponentVertexAttribute: GLenum;
    normalized: boolean;
    stride: number;
    offset: number;
    constructor(attributeLocation: number, componentsNumberPerVertexAttribute: number, typeComponentVertexAttribute: GLenum, normalized: boolean, stride: number, offset: number);
}
//# sourceMappingURL=GLAttributeInfo.d.ts.map